Beginnings: a table with some stuff in it; can select cells
I just discovered https://codereview.qt-project.org/#/c/105634/
and https://codereview.qt-project.org/#/c/104887/ today, which
made selection possible in the old Controls 1 TreeView.  It turns
out that ItemSelectionModel might be usable here too.  But the
problem is making some use of the onSelectionChanged signal in QML.

Hard-coding data in QML is not cool: the idea is to load and store
some sort of files later on.  But it's inspired by
qtbase/examples/widgets/itemviews/spreadsheet for now.