beginnings of formula evaluation; load file from command line
./spreadsheet.qml expenses.json
will run the spreadsheet and load that file.

A cell can have a formula like "v(2) * v(4)"
which means multiply the value from col 2 of the current row
by the value from col 4 of the current row.  The two-argument
form v(col, row) can grab data from other cells.  But I guess
the evaluation should be recursive, actually, in case the other
cell has a formula... this is not yet done.

Now we depend on Q_INVOKABLE QVariant QQmlTableModel::data()
to avoid needing to redo the work to get a displayable value
(e.g. the currency amounts are objects and we don't want to have
to look for a "display" role inside the object).  data() already
does this for us so we need to be able to call it.