use TableModel display/editRoleProvider to get the right stuff both ways
Exporting markdown was exporting formulas because DisplayRole was
returning them.  I want to export the evaluated result for DisplayRole,
but the evaluation has to be in JS code; therefore, the model must
always invoke some JS code when preparing a string to return for
the display role, so that the formula evaluation can be done.
(Or else the model could store display and edit strings separately,
but the model we started with had only the formulas in the last column.)

This makes it even more clear that JSON is our storage format, and
markdown is only meant for displaying calculated results (inserting
into wikis and such things).